# framework/gin/example/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/del/{id}": {
            "delete": {
                "summary": "删除",
                "responses": {}
            }
        },
        "/v1/person/": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "获取人物",
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "person"
                ],
                "summary": "添加",
                "parameters": [
                    {
                        "description": "人物添加req",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/person/{id}": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "通过id获取人物",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Person UserId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "summary": "更新人物",
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.Person": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "userName": {
                    "type": "string"
                }
            }
        }
    }
}